+++ agent/src/main/java/com/teamscale/jacoco/agent/AgentBase.java
@@ -1,7 +1,9 @@
 package com.teamscale.jacoco.agent;
 
+import com.teamscale.client.ProxySystemProperties;
 import com.teamscale.jacoco.agent.options.AgentOptions;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
+import org.conqat.lib.commons.filesystem.FileSystemUtils;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
 import org.eclipse.jetty.servlet.ServletContextHandler;
@@ -12,7 +14,9 @@ import org.glassfish.jersey.servlet.ServletContainer;
 import org.jacoco.agent.rt.RT;
 import org.slf4j.Logger;
 
+import java.io.IOException;
 import java.lang.management.ManagementFactory;
+import java.nio.file.Path;
 
 /**
  * Base class for agent implementations. Handles logger shutdown, store creation and instantiation of the
@@ -36,6 +40,7 @@ public abstract class AgentBase {
 	/** Constructor. */
 	public AgentBase(AgentOptions options) throws IllegalStateException {
 		this.options = options;
+		setProxyPasswordFromFile(options.getProxyPasswordPath());
 		try {
 			controller = new JacocoRuntimeController(RT.getAgent());
 		} catch (IllegalStateException e) {
@@ -53,6 +58,23 @@ public abstract class AgentBase {
 				throw new IllegalStateException("Control server not started.", e);
 			}
 		}
+	}
+
+	/** Sets the proxy password JVM property from a file for both http and https. */
+	private void setProxyPasswordFromFile(Path proxyPasswordFilePath) {
+		if (proxyPasswordFilePath == null) {
+			return;
+		}
+		try {
+			String proxyPassword = FileSystemUtils.readFileUTF8(proxyPasswordFilePath.toFile()).trim();
+			new ProxySystemProperties(ProxySystemProperties.Protocol.HTTP).setProxyPassword(proxyPassword);
+			new ProxySystemProperties(ProxySystemProperties.Protocol.HTTPS).setProxyPassword(proxyPassword);
+		} catch (IOException e) {
+			logger.error(
+					"Unable to open file containing proxy password. Please make sure the file exists and the user has the permissions to read the file.",
+					e);
+		}
+
 	}
 
 	/**